+++ src/main/java/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 public class Calculator {
