+++ src/main/java/calculator/CalculatorMain.java
@@ -1,6 +1,4 @@
-package caculator;
-
-import calculator.CalculatorController;
+package calculator;
 
 public class CalculatorMain {
     public static void main(String[] args) {
