+++ src/main/java/calculator/Input.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import java.util.List;
