+++ src/main/java/calculator/InputView.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 import java.util.Scanner;
