+++ src/main/java/calculator/OutputView.java
@@ -1,4 +1,4 @@
-package caculator;
+package calculator;
 
 public class OutputView {
