+++ src/test/java/calculator/CalculatorTest.java
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import caculator.Calculator;
-
 class CalculatorTest {
 
     @Test
