+++ src/test/java/calculator/InputTest.java
@@ -8,8 +8,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import caculator.Input;
-
 class InputTest {
 
     @ParameterizedTest
